+++ api/internal/gateway/user/v1/request/address.go
@@ -4,7 +4,7 @@ type CreateAddressRequest struct {
 	Lastname     string `json:"lastname,omitempty"`     // 姓
 	Firstname    string `json:"firstname,omitempty"`    // 名
 	PostalCode   string `json:"postalCode,omitempty"`   // 郵便番号
-	Prefecture   int64  `json:"prefecture,omitempty"`   // 都道府県
+	Prefecture   int32  `json:"prefecture,omitempty"`   // 都道府県
 	City         string `json:"city,omitempty"`         // 市区町村
 	AddressLine1 string `json:"addressLine1,omitempty"` // 町名・番地
 	AddressLine2 string `json:"addressLine2,omitempty"` // ビル名・号室など
@@ -16,7 +16,7 @@ type UpdateAddressRequest struct {
 	Lastname     string `json:"lastname,omitempty"`     // 姓
 	Firstname    string `json:"firstname,omitempty"`    // 名
 	PostalCode   string `json:"postalCode,omitempty"`   // 郵便番号
-	Prefecture   int64  `json:"prefecture,omitempty"`   // 都道府県
+	Prefecture   int32  `json:"prefecture,omitempty"`   // 都道府県
 	City         string `json:"city,omitempty"`         // 市区町村
 	AddressLine1 string `json:"addressLine1,omitempty"` // 町名・番地
 	AddressLine2 string `json:"addressLine2,omitempty"` // ビル名・号室など
